Fix some warnings

+++ src/PCRE.NET/PcrePatternInfo.cs
@@ -35,9 +35,9 @@ namespace PCRE
         public bool HasBackslashC => _re.GetInfoUInt32(PcreConstants.INFO_HASBACKSLASHC) != 0;
         public bool HasCrOrLf => _re.GetInfoUInt32(PcreConstants.INFO_HASCRORLF) != 0;
         public bool JChanged => _re.GetInfoUInt32(PcreConstants.INFO_JCHANGED) != 0;
-        public ulong FrameSize => (ulong)_re.GetInfoNativeInt(PcreConstants.INFO_FRAMESIZE);
-        public ulong JitSize => (ulong)_re.GetInfoNativeInt(PcreConstants.INFO_JITSIZE);
-        public ulong PatternSize => (ulong)_re.GetInfoNativeInt(PcreConstants.INFO_SIZE);
+        public ulong FrameSize => _re.GetInfoNativeInt(PcreConstants.INFO_FRAMESIZE);
+        public ulong JitSize => _re.GetInfoNativeInt(PcreConstants.INFO_JITSIZE);
+        public ulong PatternSize => _re.GetInfoNativeInt(PcreConstants.INFO_SIZE);
         public uint MatchLimit => _re.GetInfoUInt32(PcreConstants.INFO_MATCHLIMIT);
         public uint HeapLimit => _re.GetInfoUInt32(PcreConstants.INFO_HEAPLIMIT);
         public uint FirstCodeType => _re.GetInfoUInt32(PcreConstants.INFO_FIRSTCODETYPE);
